+++ src/Controller/BookmarksController.php
@@ -85,7 +85,7 @@ class BookmarksController extends AppController
         $bookmark = $this->Bookmark->get($id);
 
         if ($this->Bookmark->delete($bookmark)) {
-            $this->Flash->success(__('The bookmark %d hasbeen deleted.', $bookmark->id));
+            $this->Flash->success(__('The bookmark %d has been deleted.', $bookmark->id));
         } else {
             $this->Flash->error(__('The bookmark could not be deleted.'));
         }